refactor(selector4nix-web): extract HTTP API and web UI to a separate crate
