+++ crates/selector4nix-core/tests/integration/main.rs
@@ -5,4 +5,3 @@ mod config_test;
 mod credential_test;
 mod nar_file_service_test;
 mod nar_info_service_test;
-mod status_query_usecase_test;
