+++ crates/selector4nix-web/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "selector4nix-web"
+version = "0.0.0"
+edition = "2024"
+
+[dependencies]
+selector4nix-core = { path = "../selector4nix-core" }
+
+anyhow = { workspace = true }
+axum = { workspace = true }
+bytes = { workspace = true }
+futures = { workspace = true }
+getset = { workspace = true }
+http = { workspace = true }
+serde = { workspace = true }
+snafu = { workspace = true }
+toml = { workspace = true }
+url = { workspace = true }
