+++ crates/selector4nix-web/src/lib.rs
@@ -2,7 +2,6 @@ pub mod handlers;
 
 mod error;
 mod router;
-mod state;
 
+pub use error::WebAppError;
 pub use router::build_router;
-pub use state::AppContext;
