+++ crates/selector4nix-web/src/router.rs
@@ -2,15 +2,15 @@ use std::sync::Arc;
 
 use axum::Router;
 use axum::routing::get;
+use selector4nix_core::AppContext;
 
-use crate::api::handlers::cache_info::get_nix_cache_info;
-use crate::api::handlers::health::get_health;
-use crate::api::handlers::index::get_index;
-use crate::api::handlers::nar::get_nar;
-use crate::api::handlers::nar_info::get_nar_info;
-use crate::api::handlers::status::get_status;
-use crate::api::handlers::substituter::get_available_substituters;
-use crate::api::state::AppContext;
+use crate::handlers::cache_info::get_nix_cache_info;
+use crate::handlers::health::get_health;
+use crate::handlers::index::get_index;
+use crate::handlers::nar::get_nar;
+use crate::handlers::nar_info::get_nar_info;
+use crate::handlers::status::get_status;
+use crate::handlers::substituter::get_available_substituters;
 
 pub fn build_router(ctx: Arc<AppContext>) -> Router {
     Router::new()
