+++ crates/selector4nix/src/main.rs
@@ -8,8 +8,9 @@ use anyhow::Result as AnyhowResult;
 use clap::Parser;
 use tokio::net::TcpListener;
 
-use selector4nix_core::api::{AppContext, build_router};
+use selector4nix_core::AppContext;
 use selector4nix_core::infrastructure::config::{AppConfiguration, AppCredential};
+use selector4nix_web::build_router;
 
 use crate::cli::{Cli, Commands};
